two helpful functions
dt.appendProperty is a callback that can be used like this:
'ajax' => [
    'data' => $this->DataTables->callback('dt.appendProperty',
['fullSearch', [
        'selector' => "#$tableid-fullsearch",
        'property' => 'checked',
        'default' => $fullSearch
    ]]),
],

In this example, a "fullSearch" checkbox is placed in the document and
its state will be propagated to the server on every request.

dt.h is a helper for render functions. Use it to sanitize strings just
as the default text renderer does