Fix SQL injection vulnerability
Do not read column definitions from client (which included field names
passed to the database). Instead pass them in the controller.

To make ordering/filtering capabilities of DataTables work, find() needs
to be called with the column definition array as a fourth argument.
It is not needed for the use of DataTables without ordering/filtering or
without server-side processing (AJAX fetches coming from DataTables).