Translation for DataTables order option
This allows to pass field names or column identifiers in the order
option instead of mere column indices. Furthermore, it understands the
simple [$field => 'asc'/'desc'] style of CakePHP and translates it
accordingly.

This allows to set the order parameter in the controller, and pass it,
through a view variable, to the DataTables helper in the end.